Firmware review pass: bus concurrency, buffer and lifecycle fixes
Serialise Bus behind a recursive mutex. Both buses share UART_NUM_1 and the
HTTP handlers drive the same Bus objects as the periodic poll in the main
task, so a poll on one bus could call uart_driver_delete() while the other
was inside uart_read_bytes(). Settings sit behind a second, short-held mutex
(ordered after the bus lock) so reading or storing a value never queues
behind a multi-second RS-485 transaction.

Defer power changes to the main loop. POST /api/bus/N/power now queues the
request and returns at once instead of blocking httpd for the whole discovery
and warm-up sequence; poll() services it, and the status reported to the
client already reflects the queued request. poll() also absorbs the auto
shut-off check that used to live in wifi_controller_loop().

Replace getRepellers() with a lock-free repeller_count(). The old accessor
handed out a reference to a list the bus task mutates, so status readers on
other tasks walked it unsynchronised.

Stop transmit() powering the bus as a side effect. It called activate(), so a
colour or brightness request against an idle bus switched the rail on and left
it energised with nothing running on it.

Other correctness fixes:
- Repeller::serial was char[16] for two 8-character halves plus NUL, silently
  dropping the last digit; serial parts now stop at the first non-printable
  byte instead of substituting a placeholder for the NUL padding.
- Cartridge runtime was only persisted on a graceful power-off, so any reset
  discarded the running session. poll() now flushes it every five minutes.
- powerdown() released UART ownership without deleting the driver, so the next
  activate() skipped the delete and re-installed over a live driver.
- BUS_ERROR was cleared by the powerdown() call in activate(), letting a
  misconfigured bus retry and drive GPIO -1.
- Repellers that stopped answering were carried forward as present and warmed
  up; discovery now marks known devices OFFLINE until they respond.
- warm_up_all() with nothing on the bus parked it in BUS_WARMING_UP until auto
  shut-off; powerOn() now powers the rail back down when discovery finds
  nothing.
- receive_packet() kept its parse state in function statics shared by both bus
  instances, and polled at 100ms against an 8ms frame gap.
- load_settings() ignored every fread result, leaving a mix of defaults and
  garbage on a short file.
- heartbeat_poll() reported "all active" for an empty repeller list.
- packet_name_buffer[32] truncated tx_led_brightness_startup names at 34 bytes.
- Uninitialised Repeller::turned_on_at, unused any_active/heartbeat locals,
  getGuid() format specifiers, size_t format mismatch in sniffer_mode, and the
  -Wreorder warning in WiFiRepellerDevice.

src/CMakeLists.txt globbed src/*.*, pulling headers, idf_component.yml and
CMakeLists.txt itself into SRCS. Drop the BUILDING.md and CLAUDE.md references
left dangling by the Zigbee removal and the sdkconfig cleanup.

Builds with zero warnings under -Wall -Wextra in all three mode
configurations (WIFI_CONTROLLER, CONTROLLER, SNIFFER).

+++ src/sniffer_mode.cpp
@@ -6,6 +6,7 @@
 #include "driver/uart.h"
 #include "freertos/FreeRTOS.h"
 #include "freertos/task.h"
+#include <cinttypes>
 
 static const char* TAG = "sniffer";
 
@@ -48,7 +49,7 @@ static void processPacket() {
       for(size_t i = 0; i < buffer_index; i++) {
         snprintf(&hex_buf[i * 3], 4, "%02X ", rx_buffer[i]);
       }
-      ESP_LOGI(TAG, "[%08lu] RX: %s(PARTIAL) [%d bytes]", millis_now(), hex_buf, buffer_index);
+      ESP_LOGI(TAG, "[%08" PRIu32 "] RX: %s(PARTIAL) [%u bytes]", millis_now(), hex_buf, (unsigned)buffer_index);
     }
     buffer_index = 0;
   }
